Add `encode_action` and `decode_action`

+++ src/board/pieces/mod.rs
@@ -10,7 +10,7 @@ const NORMAL_MOVE: usize = 0;
 pub trait Piece {
     fn get_piece_symbol(&self) -> PieceSymbol;
 
-    fn format_info(&self, board: &Board, info: u32) -> String {
+    fn format_info(&self, board: &Board, info: usize) -> String {
         "".to_string()
     }
     fn parse_info(&self, board: &Board, info: String) -> u32 {
